ai-service/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.17
pydantic>=2.10.0
pydantic-settings>=2.6.0

# OpenAI
openai>=1.50.0

# Anthropic (Claude)
anthropic>=0.34.0

# LangChain (para agentes)
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Database
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.30

# Supabase (Vector Store) - opcional
supabase>=2.10.0

# Redis
redis>=5.0.0

# Utils
python-dotenv>=1.0.0
httpx>=0.27.0
tenacity>=9.0.0
nest-asyncio>=1.6.0

# Logging
structlog>=24.0.0

# Tiktoken para contar tokens
tiktoken>=0.8.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Machine Learning
torch>=2.0.0
numpy>=1.24.0

# Scheduling
apscheduler>=3.10.0
